Put all classes into one file to avoid importing problem between game and game_backend
