--- utility/flame.py
@@ -1,8 +1,0 @@
-class Flame(object):
-
-    def __init__(self, alliance=None, position=None):
-        self.alliance = alliance
-        self.position = position
-
-    def to_string(self):
-        return '0'
