+++ utility/game_backend.py
@@ -1,8 +1,4 @@
-from utility.board import Board
-from utility.null import Null
-from utility.flame import Flame
-
-class Game(object):
+class Game:
 
     def __init__(self):
         self.board = Board()
@@ -79,3 +75,105 @@ class Game(object):
                     self.board.print_board()
 
             self.make_play(p, m, s)
+
+class Flame:
+
+    def __init__(self, alliance=None, position=None):
+        self.alliance = alliance
+        self.position = position
+
+    def to_string(self):
+        return '0'
+
+class Board:
+    def __init__(self):
+        self.possible_steps = [[-1,-1], [-1,0], [-1,1], [0,-1], [0,1], [1,-1], [1,0], [1,1]]
+        self.game_tiles = []
+
+        # Fill board with nulls
+        for tile in range(100):
+            self.game_tiles.append(Null(None, None))
+
+        # Add tribe 1
+        self.game_tiles[3] = Warrior(1, 3)
+        self.game_tiles[6] = Warrior(1, 6)
+        self.game_tiles[30] = Warrior(1, 30)
+        self.game_tiles[39] = Warrior(1, 39)
+
+        # Add tribe 2
+        self.game_tiles[60] = Warrior(2, 60)
+        self.game_tiles[69] = Warrior(2, 69)
+        self.game_tiles[93] = Warrior(2, 93)
+        self.game_tiles[96] = Warrior(2, 96)
+
+
+    def print_board(self):
+        for tiles in range(100):
+            print('|', end=self.game_tiles[tiles].to_string())
+            if (tiles+1)%10 == 0:
+                print('|', end='\n')
+
+class Null:
+
+    def __init__(self, alliance=None, position=None):
+        self.alliance = alliance
+        self.position = position
+
+    def to_string(self):
+        return "-"
+
+class Warrior:
+
+    def __init__(self, alliance, position):
+        self.alliance       = alliance
+        self.position       = position
+
+    def to_string(self):
+        return "1" if self.alliance==1 else "2"
+
+    def find_moves(self, board):
+        """
+        Find possible move locations
+        """
+        moves = []
+        for step in board.possible_steps:
+            pos = self.position
+            # Convert to cartesian
+            x = pos%10
+            y = pos//10
+            while (0 <= x <= 9) and (0 <= y <= 9):
+                x += step[0]
+                y += step[1]
+                if (0 <= x <= 9) and (0 <= y <= 9):
+                    pos2 = x + 10*y
+                    if board.game_tiles[pos2].to_string() == '-':
+                        moves.append(pos2)
+                    else:
+                        break
+                else:
+                    break
+        return moves
+
+    def find_shoots(self, board, move):
+        """
+        Find possible shoot locations from the position move.
+        """
+        old_position = self.position
+        moves = []
+        for step in board.possible_steps:
+            pos = move
+            # Convert to cartesian
+            x = pos%10
+            y = pos//10
+            while (0 <= x <= 9) and (0 <= y <= 9):
+                x += step[0]
+                y += step[1]
+                if (0 <= x <= 9) and (0 <= y <= 9):
+                    pos2 = x + 10*y
+                    if board.game_tiles[pos2].to_string() == '-' or pos2 == old_position:
+                        moves.append(pos2)
+                    else:
+                        break
+                else:
+                    break
+        return moves
